pyproject.toml
[tool.poetry]
name = "blog-drf"
version = "0.1.0"
description = "full rest api blog project"
authors = ["orhan <orhancansu98@gmail.com>",]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.4.1"
attrs = "21.2.0"
autopep8 = "1.6.0"
certifi = "2021.10.8"
charset-normalizer = "2.0.7"
click = "8.0.3"
coreapi = "2.3.3"
coreschema = "0.0.4"
Django = "3.2.9"
django-ckeditor = "6.1.0"
django-js-asset = "1.2.2"
django-rest-swagger = "2.2.0"
django-taggit = "2.0.0"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "5.0.0"
drf-spectacular = "0.21.0"
idna = "3.3"
inflection = "0.5.1"
itypes = "1.2.0"
Jinja2 = "3.0.3"
jsonschema = "4.2.1"
MarkupSafe = "2.0.1"
mypy-extensions = "0.4.3"
openapi-codec = "1.3.2"
pathspec = "0.9.0"
Pillow = "8.4.0"
platformdirs = "2.4.0"
pycodestyle = "2.8.0"
PyJWT = "2.3.0"
pyrsistent = "0.18.0"
pytz = "2021.3"
PyYAML = "6.0"
regex = "2021.11.10"
requests = "2.26.0"
simplejson = "3.17.6"
sqlparse = "0.4.2"
toml = "0.10.2"
tomli = "1.2.2"
typing-extensions = "4.0.0"
uritemplate = "4.1.1"
urllib3 = "1.26.7"

[tool.poetry.dev-dependencies]
black = "21.11b1"
isort = "5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
